+++ src/Modelo/Sector.java
@@ -1,30 +1,30 @@
 /**
- ** NOMBRE CLASE: 
+ ** NOMBRE CLASE:
  **	  Sector.java
  **
  ** DESCRIPCION:
  **       Abstracción de un Sector
- **       
+ **
  **
  ** DESARROLLADO POR:
  **       Juan Antonio Aranda Ortega
- **	   
- **        
+ **
+ **
  ** SUPERVISADO POR:
- **       
+ **
  **
  ** HISTORIA:
- ** 	000 - Mayo 22, 2012 - Creacion 
+ ** 	000 - Mayo 22, 2012 - Creacion
  **
  ** NOTAS:
- **   
+ **
  **
  */
 package Modelo;
 
 public class Sector {
-    
-    private int OID;
+
+    private Long OID;
     private String descripcion;
 
     /**
@@ -36,19 +36,19 @@ public class Sector {
     /**
      * Constructor con todos los atributos
      * @param OID
-     * @param descripcion 
+     * @param descripcion
      */
-    public Sector(int OID, String descripcion) {
+    public Sector(Long OID, String descripcion) {
         this.OID = OID;
         this.descripcion = descripcion;
     }
 
     //Métodos set y get
-    public int getOID() {
+    public Long getOID() {
         return OID;
     }
 
-    public void setOID(int OID) {
+    public void setOID(Long OID) {
         this.OID = OID;
     }
 
@@ -59,5 +59,5 @@ public class Sector {
     public void setDescripcion(String descripcion) {
         this.descripcion = descripcion;
     }
-    
+
 }
